ruleset.uc: don't trim newline before comment sign
When flow offloading is enabled, the comment block is inserted in the
ruleset like this:

table inet fw4 {	#
	# Flowtable
	#

This is due to the trimming of newlines which was done to avoid having
to modify all the test files. A better solution is to just add an extra
newline in the template.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>
Reviewed-by: Jo-Philipp Wich <jo@mein.io>